requirements.txt
# DeepSeek OCR + PageIndex System
# Minimal requirements - vLLM and cuda packages should be installed separately via uv

# Core OCR & API
openai>=1.0.0           # OpenAI client for vLLM
httpx>=0.27.0           # HTTP client  
fastapi>=0.100.0        # API framework
uvicorn>=0.20.0         # ASGI server

# Document Processing
PyMuPDF>=1.26.0         # PDF rendering
PyPDF2>=3.0.0           # PDF reading
Pillow>=10.0.0          # Image processing

# Database
SQLAlchemy>=2.0.0       # ORM

# Configuration
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# PageIndex Dependencies
tiktoken>=0.5.0         # Token counting
PyYAML>=6.0.0           # Config files

# Optional: Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
